Use ES6 modules, add Google Calendar `RRULE` parsing.
This commit changes the structure of the JavaScript to use proper ES6
Modules native to Web browsers. There is still no build system in use
but this moves us closer to that if necessary. Meanwhile, we make use
of the `ical-expander` NPM package to parse iCalendar's `RRULE`
properties for the fetched Google Calendar `.ics` files. This fixes a
bug where only the first occurrence of a sequence was displayed.